+++ packages/games/src/lib/common/controller.tsx
@@ -2,15 +2,15 @@ import * as React from 'react';
 import { useFormContext } from 'react-hook-form';
 
 import { CDN_URL } from '../constants';
+import useMediaQuery from '../hooks/use-media-query';
 import { Button } from '../ui/button';
 import { FormControl, FormField, FormItem, FormLabel, FormMessage } from '../ui/form';
 import { cn } from '../utils/style';
 import { BetCount } from './bet-count';
 import { BetCountSlider } from './containers';
+import { IncreaseByInput } from './increase-by-input';
 import { StopGainLossInput } from './stop-gain-loss-input';
 import { WagerBalance, WagerCurrency, WagerInput, WagerSetterButtons } from './wager';
-import { IncreaseByInput } from './increase-by-input';
-import useMediaQuery from '../hooks/use-media-query';
 
 interface WagerFormFieldProps {
   customLabel?: string;
@@ -27,7 +27,14 @@ interface Props {
 
 export const BetControllerTitle: React.FC<Props> = ({ children, className }) => {
   return (
-    <h1 className={cn('wr-text-lg wr-font-bold  lg:wr-flex wr-hidden', className)}>{children}</h1>
+    <h1
+      className={cn(
+        'wr-text-lg wr-font-bold  lg:wr-flex wr-hidden bet-controller-title',
+        className
+      )}
+    >
+      {children}
+    </h1>
   );
 };
 
@@ -45,7 +52,7 @@ export const BetCountFormField: React.FC<{
         name="betCount"
         render={({ field }) => (
           <FormItem
-            className={cn('wr-mb-3 lg:wr-mb-6', {
+            className={cn('wr-mb-3 lg:wr-mb-6 bet-count', {
               'wr-hidden lg:!wr-block': hideSm,
             })}
           >
@@ -80,7 +87,7 @@ export const AutoBetCountFormField: React.FC<{
         control={form.control}
         name="betCount"
         render={({ field }) => (
-          <FormItem className={cn('wr-mb-3')}>
+          <FormItem className={cn('wr-mb-3 autobet-form-field')}>
             <FormLabel>{isMobile ? '#' : 'Number'} of Bets</FormLabel>
 
             <FormControl>
@@ -117,7 +124,7 @@ export const AutoBetStopGainFormField = ({
       control={form.control}
       name="stopGain"
       render={({ field }) => (
-        <FormItem className="lg:wr-mb-3 wr-mb-0 wr-w-full">
+        <FormItem className="lg:wr-mb-3 wr-mb-0 wr-w-full autobet-stop-gain">
           <FormLabel className={cn(labelClassName)}>Stop on Profit</FormLabel>
           <FormControl>
             <StopGainLossInput
@@ -150,7 +157,7 @@ export const AutoBetStopLossFormField = ({
       control={form.control}
       name="stopLoss"
       render={({ field }) => (
-        <FormItem className="lg:wr-mb-3 wr-mb-0 wr-w-full">
+        <FormItem className="lg:wr-mb-3 wr-mb-0 wr-w-full autobet-stop-loss">
           <FormLabel className={cn(labelClassName)}>Stop on Loss</FormLabel>
           <FormControl>
             <StopGainLossInput
@@ -186,7 +193,7 @@ export const AutoBetIncreaseOnWin = ({
       control={form.control}
       name="increaseOnWin"
       render={({ field }) => (
-        <FormItem className="wr-mb-3">
+        <FormItem className="wr-mb-3 autobet-increase-on-win">
           <FormLabel className={cn(labelClassName)}>On Win</FormLabel>
           <FormControl>
             <div className="wr-flex wr-w-full wr-bg-zinc-800 wr-rounded-md wr-py-0.5 wr-px-2 wr-pr-0.5 wr-items-center">
@@ -245,7 +252,7 @@ export const AutoBetIncreaseOnLoss = ({
       control={form.control}
       name="increaseOnLoss"
       render={({ field }) => (
-        <FormItem className="wr-mb-3">
+        <FormItem className="wr-mb-3 autobet-increase-on-loss">
           <FormLabel className={cn(labelClassName)}>On Loss</FormLabel>
           <FormControl>
             <div className="wr-flex wr-w-full wr-bg-zinc-800 wr-rounded-md wr-py-0.5 wr-px-2 wr-pr-0.5 wr-items-center">
@@ -299,7 +306,7 @@ export const WagerFormField: React.FC<WagerFormFieldProps> = ({
       control={form.control}
       name="wager"
       render={({ field }) => (
-        <FormItem className={cn(className, 'wr-mb-3 lg:wr-mb-6')}>
+        <FormItem className={cn(className, 'wr-mb-3 lg:wr-mb-6 wager-input')}>
           <FormLabel className={cn('wr-leading-4 wr-mb-3 lg:wr-mb-[6px] lg:wr-leading-6')}>
             {customLabel ? customLabel : 'Wager'}
             <div>