+++ packages/tailwind-config/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @winrlabs/tailwind-config
 
+## 2.0.0
+
+### Major Changes
+
+- 0482481: (controller.tsx): reorganize imports and add IncreaseByInput component
+  (controller.tsx): refactor BetControllerTitle component for better styling
+  (controller.tsx): update styling for BetCountFormField and AutoBetCountFormField components
+  (controller.tsx): update styling for AutoBetStopGainFormField and AutoBetStopLossFormField components
+  (controller.tsx): update styling for AutoBetIncreaseOnWin and AutoBetIncreaseOnLoss components
+  (controller.tsx): update styling for WagerFormField component
+  (increase-by-input.tsx): add class 'increase-by-input' for styling
+  (pre-bet-button.tsx): add class 'submit-button' for styling
+  (wager.tsx): update styling for WagerInput and WagerSetterButtons components
+  (wager.tsx): update styling for TotalWager and MaxPayout components
+  (auto-controller.tsx): update styling for PreBetButton component
+  (index.tsx): refactor BetController component for better styling and remove backgroundColor prop
+  (manual-controller.tsx): Add CSS class 'max-payout' for styling consistency
+  (template.tsx): Remove unused backgroundColor property from options
+  (manual-controller.tsx): Add CSS class 'max-payout' for styling consistency
+  (tailwind.config.ts): Update animation keyframes to include border-bottom-width property
+  (index.tsx): Add formClassName prop to BetController component for custom styling
+
 ## 1.3.9
 
 ### Patch Changes